Error Code 변수 사용 및 Repository와 ServicService 역할 구분

+++ src/domains/connect-logs/connect-logs.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { ConnectLogsRepository } from "./connect-logs.repository";
 import { MyPaginationQuery } from "../base/pagination-query";
@@ -6,6 +6,10 @@ import { Pagination } from "nestjs-typeorm-paginate";
 import { ConnectLog } from "./entities/connect-log.entity";
 import { User } from "../users/entities/user.entity";
 import { ConnectLogListResponseDto } from "./dto/connect-log-list-response.dto";
+import {
+  CONNECT_LOG_EXCEPTION,
+  USER_EXCEPTION,
+} from "../../exception/error-code";
 
 @Injectable()
 export class ConnectLogsService {
@@ -23,11 +27,19 @@ export class ConnectLogsService {
 
   // 접속 로그 생성
   createLog(url: string, ip: string, user: User): Promise<ConnectLog> {
+    if (!user) {
+      throw new NotFoundException(USER_EXCEPTION.USER_NOT_FOUND);
+    }
+
     return this.logsRepository.createLog(url, ip, user);
   }
 
   // 접속 로그 삭제
   deleteLog(logId: number): Promise<void> {
+    if (!logId) {
+      throw new NotFoundException(CONNECT_LOG_EXCEPTION.CONNECT_LOG_NOT_FOUND);
+    }
+
     return this.logsRepository.deleteLog(logId);
   }
 }
